Create A2Q3 ~ Griffin Ehnstrom.js
following the guide lines of question 3 I added functions translate and scale to be able to use x and y coordinates to move my shape and then the size to scale the shape down so I can make multiple and they don't over lap. one is massive and in the middle of the canvas and the other is small and in the upper top left where its supposed to start according to the guidelines of the question.