Create A2Q4 ~ Griffin Ehnstrom.js
https://editor.p5js.org/gkehnstrom/sketches/WG3l341Ry

for assignment 2 question 4 we were told to choose 2 flags, one of them that started with the first letter of our name and the first letter of our last name. and then combine the two flags to create your own flag. So I combined the flag of Eritrea because my last name is Ehnstrom, and since my name is Griffin is chose the flag of Guatemala.